Adding row to section via Eureka form is now much more secure, added padding between searchBar and tableView, fixed nutrient model and spoonacular type, changed view.safeAreaLayoutGuide.bottomAnchor to view.bottomAnchor.

+++ Plot/Model/GroceryProduct.swift
@@ -34,7 +34,7 @@ struct GroceryProduct: Codable, Equatable, Hashable {
     let servings: Servings?
     let serving_size: String?
     let number_of_servings: Double?
-    let spoonacularScore: Int?
+    let spoonacularScore: Double?
     let image: String?
     var recipe: [String: Double]?
     var bool: Bool?
